First commit to repository. All that's been done is the creation of a simple framework for the project.
What works:
Instantiation and set-up of the SEA, FISH and SAILOR objects.
A simple fishing method implemented in the SEA class. Though at the
moment no fish school's population drops below 1 fish.
I have also added a slow but illuminating scatter plot to ensure that the thing
roughly works as intended. But there is ample room for improvements all
over the place.

What's left to implement: A LOT!
Probably the first things to add are movement methods for both the fish
schools and the fishermen.
There's also the matter of corner cases to take care of. Like the above
mentioned and the question of prioritization/conflict resolution since
the current implementation allows multiple fishermen to target/interact
with the same fish schools if they are close to eachother.